refactor page and add some team

+++ repository/db/dao/race.go
@@ -50,8 +50,8 @@ func (dao *RaceDao) GetRaceCount() (count int64, err error) {
 	return count, nil
 }
 
-func (dao *RaceDao) FindRaceByPage(page int, perPage int) (races []model.Race, err error) {
-	if err = dao.DB.Model(&model.Race{}).Offset((page - 1) * perPage).Limit(perPage).Find(&races).Error; err != nil {
+func (dao *RaceDao) FindRaceByPage(page int32, perPage int32) (races []model.Race, err error) {
+	if err = dao.DB.Model(&model.Race{}).Offset(int((page - 1) * perPage)).Limit(int(perPage)).Find(&races).Error; err != nil {
 		return nil, err
 	}
 	return races, nil
